Time: 15 ms (77.42%), Space: 49.6 MB (52.53%) - LeetHub

+++ 0387-first-unique-character-in-a-string/0387-first-unique-character-in-a-string.java
@@ -5,7 +5,7 @@ class Solution {
         int arr[] = new int[26];
         
         for(char ch : s.toCharArray())
-            arr[(int)ch-'a']+=1;
+            arr[ch-'a']+=1;
         
         for(int i=0; i<s.length(); i++)
             if(arr[(int)s.charAt(i)-'a'] == 1) return i;
